requirements.txt
alembic==1.7.1
amqp==5.0.6
attrs==21.2.0
Babel==2.9.1
bcrypt==3.2.0
billiard==3.6.4.0
black==21.9b0
blinker==1.4
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flower==1.0.0
greenlet==1.1.1
humanize==3.11.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
multitasking==0.0.9
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.3.0
pluggy==1.0.0
pmaw==2.0.0
praw==7.4.0
prawcore==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
toml==0.10.2
tomli==1.2.1
tornado==6.1
tqdm==4.62.3
tweepy==4.0.0
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.6
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.1
WTForms==2.3.3
yahooquery==2.2.15
yfinance==0.1.63
zxcvbn==4.4.28
